More cleanup of websocket protocol
