+++ src/gateway/util/websocket.h
@@ -1,6 +1,7 @@
 #ifndef websocket_h
 #define websocket_h
 
+#include <iomanip>
 #include <iostream>
 #include <map>
 #include <openssl/bio.h>
@@ -8,11 +9,9 @@
 #include <openssl/evp.h>
 #include <openssl/sha.h>
 #include <spdlog/spdlog.h>
-#include <stdexcept>
 #include <sstream>
+#include <stdexcept>
 #include <string>
-#include <iostream>
-#include <iomanip>
 
 using namespace std;
 
@@ -20,20 +19,24 @@ using namespace std;
 static const string ws_magic_string = "258EAFA5-E914-47DA-95CA-C5AB0DC85B11";
 static const string ws_request_header = "Sec-WebSocket-Key";
 
-static const string ws_response =
-  "HTTP/1.1 101 Switching Protocols\r\n"                                       \
-  "Upgrade: websocket\r\n"                                                     \
-  "Connection: Upgrade\r\n"                                                    \
-  "Sec-WebSocket-Accept: ";
+static const string ws_response = "HTTP/1.1 101 Switching Protocols\r\n"
+                                  "Upgrade: websocket\r\n"
+                                  "Connection: Upgrade\r\n"
+                                  "Sec-WebSocket-Accept: ";
 
 #define WS_ACCEPT_RESPONSE_LEN 130
 #define SHA1_HUMAN_LEN 41
 
-string base64_encode(const string &original);
+string base64_encode(unsigned char *original);
 string base64_decode(const string &encoded);
 string sha1(const string &input);
 map<string, string> parse_http_headers(const string &headers);
 string create_websocket_response_nonce(const string &websocket_request_key);
 string websocket_request_response(const string &client_http_request);
+
+void printStringAsHex(const string &str);
+void printByteAsHex(const string &message, const uint8_t byte);
+bool decodeWebSocketFrame(string frame, string &message);
+vector<uint8_t> encodeWebsocketFrame(const string &message);
 
 #endif
